refactor: API 배럴 파일 사용수정
