--- frontend/src/api/get/index.ts
@@ -1,2 +1,0 @@
-export * from './MainPage';
-export * from './ProjectPage';
