+++ frontend/src/hook/api/useGroupNames.ts
@@ -1,4 +1,4 @@
-import { getGroupNames } from '@api/get';
+import { getGroupNames } from '@api/get/ProjectPage';
 import { useQuery } from '@tanstack/react-query';
 
 export default function useGroupNames(generation: string) {
