+++ frontend/src/hook/api/useProjectElapsedTime.ts
@@ -1,4 +1,4 @@
-import { getElapsedTime } from '@api/get';
+import { getElapsedTime } from '@api/get/ProjectPage';
 import { useSuspenseQuery } from '@tanstack/react-query';
 
 export default function useProjectElapsedTime(project: string) {
