+++ frontend/src/hook/api/useRankings.ts
@@ -1,4 +1,4 @@
-import { getRankings } from '@api/get';
+import { getRankings } from '@api/get/MainPage';
 import { useSuspenseQuery } from '@tanstack/react-query';
 
 export default function useRankings(generation: string) {
