+++ frontend/src/hook/api/useTop5ResponseTime.ts
@@ -1,4 +1,4 @@
-import { getTop5ResponseTime } from '@api/get';
+import { getTop5ResponseTime } from '@api/get/MainPage';
 import { useSuspenseQuery } from '@tanstack/react-query';
 
 export default function useTop5ResponseTime(generation: string) {
