+++ frontend/src/hook/api/useTop5Traffic.ts
@@ -1,4 +1,4 @@
-import { getTop5Traffic } from '@api/get';
+import { getTop5Traffic } from '@api/get/MainPage';
 import { useSuspenseQuery } from '@tanstack/react-query';
 
 export default function useTop5Traffic(generation: string) {
